refactor: globe 데이터 포인트를 globeData.ts로 분리
